+++ src/challenges/js/30DaysJs-31/src/Subscriber.js
@@ -1,3 +1,3 @@
-export class Newsletter {
+export class Subscriber {
   // Tu código aquí 👈
 }
